+++ src/onegov.winterthur/onegov/winterthur/assets/js/winterthur.js
@@ -1,3 +1,4 @@
+// custom map for winterthur
 function spawnDefaultMap(target, options) {
     var service = 'https://stadtplan.winterthur.ch/wms/Hintergrundkarte_LK_AV_Situationsplan?';
     var layers = 'Hintergrundkarte_LK_AV_Situationsplan';
@@ -9,10 +10,19 @@ function spawnDefaultMap(target, options) {
     return map
 }
 
+// replace checkmarks in directory entries
 $(document).ready(function() {
     $('.field-display dd').html(function() {
         return $(this).html().split('<br>').map(function(html) {
             return html.replace(/\s*[✓\-]{1}/, ' •');
         }).join('<br>');
+    });
+});
+
+// add hover clicks to directory entries
+$(document).ready(function() {
+    $('.with-lead li').click(function(e) {
+        window.location = $(this).find('a').attr('href');
+        e.preventDefault();
     });
 });